+++ CHANGES.md
@@ -8,6 +8,8 @@ To be released.
 
 ### Backward-incompatible API changes
 
+ -  Added `StateCompleterSet<T>.ComplementAll` property.  [[#1358], [#1386]]
+ -  Added `StateCompleterSet<T>.ComplementLatest` property.  [[#1358], [#1386]]
  -  `BlockPerception` now implements `IBlockExcerpt` interface.  [[#1440]]
      -  `BlockPerception.Excerpt` property removed.
  -  `TotalDifficultyComparer` now implements `IComparer<IBlockExcerpt>`
@@ -57,6 +59,9 @@ To be released.
 
 ### Behavioral changes
 
+ -  `StateCompleterSet<T>.Recalculate` now evaluates states even for those
+    already in `IStateStore`.  Moreover, it also terminates early if possible
+    after reaching the `BlockHash` provided with a call.  [[#1358], [#1386]]
  -  `TotalDifficultyComparer` no longer considers perceived time when comparing
     `IBlockExcerpt`s.  [[#1442]]
  -  General logic for determining the canonical chain has been updated.
@@ -78,6 +83,8 @@ To be released.
 
 ### CLI tools
 
+[#1358]: https://github.com/planetarium/libplanet/issues/1358
+[#1386]: https://github.com/planetarium/libplanet/pull/1386
 [#1435]: https://github.com/planetarium/libplanet/issues/1435
 [#1440]: https://github.com/planetarium/libplanet/pull/1440
 [#1442]: https://github.com/planetarium/libplanet/pull/1442